requirements.txt
Django>=4.2.0
openai>=1.0.0
python-dotenv>=1.0.0
openai
PyYAML
openai
PyYAML
